requirements.txt
torch>=1.9.0
numpy>=1.19.0
opencv-python>=4.5.0
pandas>=1.3.0
tqdm>=4.60.0
pyyaml>=5.4.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
